LATTICE_3/G — the patch index buffers are uint16
Every index the three patch IB builders emit is a vertex ordinal in the
unified decode's address space, whose top is UG_DECODE_VERTS = 14,977. It
was never a 32-bit number. A static_assert in createMeshBuffers now says
so, so the width cannot silently become wrong if the decode grows:

    static_assert(Dim::UG_DECODE_VERTS - 1 <= 0xFFFFu, ...)

                    indices    before      after
    Patch IB         50,688   202,752 B  101,376 B
    Patch IB Cap-Only 26,112   104,448 B   52,224 B
    Patch IB LOD1     19,200    76,800 B   38,400 B
    TOTAL             96,000   384,000 B  192,000 B   (187.5 KiB back)

The index FETCH halves with the storage — a few MB a frame off the bus at
LOD0 instance counts, and more where the bus is shared. Zero visual change:
same indices, same counts, same draws.

The five mesh-gen families keep Uint32. Their indices are slot-based
absolute addresses into a shared VB and run well past 65,535 — the two
generic helpers (`draw_indexed_mesh`, `draw_shadow_indexed_mesh`) that
serve them are untouched. Three SetIndexBuffer calls cross to Uint16:
`draw_patch_terrain_plan_slot` (draw-plan segments A/B/C),
`draw_patch_terrain_direct` (the gallery snapshot pass), and
`draw_shadow_patch_terrain` (whose band-1 sibling reuses the same bind).

R-F, answered: THERE ARE NO BYTE OFFSETS ON ANY PATCH SetIndexBuffer. The
handoff flagged the LOD1 "clean prefix vs zoned tail" as the one place
bytes might appear — it is a COUNT (`patch_index_count_lod1_live()`
selecting RingClean_ or RingZoned_ over one buffer), not an offset. The
`FC_SEG_*_OFF` byte offsets are on the VERTEX buffer (the visible-index
list window), not the IB. And `reset_frustum_indirect` writes indexCount /
firstIndex as counts with firstIndex always 0, so the indirect path needs
no change at all.

`pad_u16` pads an odd index count with one never-drawn index, because
WriteBuffer wants a 4-byte multiple. It does not fire today — every builder
emits triangles in quad pairs, so all three counts are even — and the draw
count is captured before the pad in every case. A bound stated is a bound
that cannot rot.

The [GPU Budget] boot line reports the new sizes for free: makeBuffer
meters the size it is handed (W4).

Gates, all green at this tree: G-LAW 2, TU (3 TUs), shell, sha256, score,
WGSL, organ gap, organ ledger, mirror. Binding surface PASS but for S-6
(dirty at the time of the run).

+++ src/cartridges/the_board/realization/renderer.hpp
@@ -800,7 +800,10 @@ namespace t7 {
             ) {
                 pass.SetBindGroup(2, stateGroup);
                 pass.SetVertexBuffer(0, visibleList, visibleOffset, visibleBytes);
-                pass.SetIndexBuffer(indexBuffer, wgpu::IndexFormat::Uint32);
+                // LATTICE_3: the three patch IBs are uint16. The indirect args
+                // carry indexCount / firstIndex as COUNTS, not bytes, so
+                // reset_frustum_indirect is untouched by the width change.
+                pass.SetIndexBuffer(indexBuffer, wgpu::IndexFormat::Uint16);
                 pass.DrawIndexedIndirect(indirectArgs, indirectOffset);
             }
 
@@ -825,7 +828,7 @@ namespace t7 {
                 pass.SetBindGroup(2, stateGroup);
                 pass.SetBindGroup(3, texGroup);
                 pass.SetVertexBuffer(0, visibleList);
-                pass.SetIndexBuffer(indexBuffer, wgpu::IndexFormat::Uint32);
+                pass.SetIndexBuffer(indexBuffer, wgpu::IndexFormat::Uint16);   // LATTICE_3
                 pass.DrawIndexed(indexCount, instanceCount, 0, 0, firstInstance);
             }
 
@@ -1089,7 +1092,9 @@ namespace t7 {
                 uint32_t instanceCount
             ) {
                 pass.SetPipeline(shadowPatchTerrainPipeline_);
-                pass.SetIndexBuffer(indexBuffer, wgpu::IndexFormat::Uint32);
+                // LATTICE_3. Band 1 below reuses this bind, so both shadow
+                // bands cross to uint16 together — they always shared the IB.
+                pass.SetIndexBuffer(indexBuffer, wgpu::IndexFormat::Uint16);
                 pass.DrawIndexed(indexCount, instanceCount);
             }
